+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### To be Released
 
+* Add `env-get` command to retrieve the value of a specific environment variable [#643](https://github.com/Scalingo/cli/pull/643)
 * Error message are outputted on stderr [#639](https://github.com/Scalingo/cli/pull/639)
 * Automatically prefix the integration URL with https if none is provided [#642](https://github.com/Scalingo/cli/pull/642)
 * `backups-download` downloads the most recent backup if none is specified [#636](https://github.com/Scalingo/cli/pull/636)